Move deleteLayoutContents(...) into Utilities
